+++ src/main/java/com/idukbaduk/itseats/rider/service/RiderService.java
@@ -41,30 +41,4 @@ public class RiderService {
                 .isWorking(rider.getIsWorking())
                 .build();
     }
-
-    @Transactional
-    public void acceptDelivery(String username, Long orderId) {
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-        Rider rider = riderRepository.findByMember(member)
-                .orElseThrow(() -> new RiderException(RiderErrorCode.RIDER_NOT_FOUND));
-
-        Order order = orderRepository.findById(orderId)
-                .orElseThrow(() -> new OrderException(OrderErrorCode.ORDER_NOT_FOUND));
-
-        order.updateOrderStatusAccept(rider, OrderStatus.RIDER_READY);
-    }
-
-    @Transactional
-    public void updateOrderStatusAfterAccept(String username, Long orderId, OrderStatus orderStatus) {
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-        Rider rider = riderRepository.findByMember(member)
-                .orElseThrow(() -> new RiderException(RiderErrorCode.RIDER_NOT_FOUND));
-
-        Order order = orderRepository.findByRiderAndOrderId(rider, orderId)
-                .orElseThrow(() -> new OrderException(OrderErrorCode.ORDER_NOT_FOUND));
-
-        order.updateStatus(orderStatus);
-    }
 }
